+++ index.html
@@ -236,7 +236,7 @@
   </template>
 
   <!-- Подключаем js -->
-  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
