+++ js/util.js
@@ -1,0 +1,38 @@
+// Функция для получения рандомного числа из переданного диапазона.
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+// Функция, чтобы рандомное число не повторялось.
+const uniqueRandomInteger = (min, max) => {
+  const previousValues = [];
+
+  return function() {
+    let currentValue = getRandomInteger(min,max);
+    // Проверка, чтобы вызовы функций не превышали максимальное количество
+    if(previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    // Проверка, что число есть в хранилище
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export {getRandomInteger, uniqueRandomInteger, createIdGenerator};
